Chnages while sending the qut return back to tables by Nikhil

+++ Admin/CNCBending.aspx.cs
@@ -638,8 +638,8 @@ public partial class Admin_CNCBending : System.Web.UI.Page
                             }
                             else
                             {
-                                //Inwardqty = 0;
-                                int TotalReturn_Outward = Convert.ToInt32(hdnInwardQty.Value) - Convert.ToInt32(txtReturnInward.Text);
+                                Inwardqty = 0;
+                                //int TotalReturn_Outward = Convert.ToInt32(hdnInwardQty.Value) - Convert.ToInt32(txtReturnInward.Text);
                                 int TotalReturnInward = Convert.ToInt32(Inwardqty.ToString()) + Convert.ToInt32(txtReturnInward.Text);
 
                                 ////Updated current stage
@@ -649,7 +649,8 @@ public partial class Admin_CNCBending : System.Web.UI.Page
                                 ////Updated Prev stage 
                                 //SqlCommand cmdupdate1 = new SqlCommand("UPDATE [dbo].[tblLaserPrograming] SET [InwardQty] = '" + TotalReturnInward + "' ,[IsComplete] = NULL  WHERE SubOA='" + hdnSubOANo.Value + "'", con);
                                 //cmdupdate1.ExecuteNonQuery();
-                                Setreturnquantity(TotalReturn_Outward, TotalReturnInward);
+                                //Setreturnquantity(TotalReturn_Outward, TotalReturnInward);
+                                InsertFullReturnQuantity(TotalReturnInward);
                             }
 
 
@@ -945,8 +946,20 @@ public partial class Admin_CNCBending : System.Web.UI.Page
 
             if (Success >= 0)
             {
-                SqlCommand cmdDelete = new SqlCommand("DELETE FROM [tblCNCBending] WHERE SubOA='" + hdnSubOANo.Value + "'", con);
-                cmdDelete.ExecuteNonQuery();
+                SqlCommand cmdselect = new SqlCommand("select InwardQty from  tblCNCBending  WHERE SubOA='" + hdnSubOANo.Value + "'", con);
+                Object Inwardqty = cmdselect.ExecuteScalar();
+                if (Inwardqty == null)
+                {
+                    SqlCommand cmdDelete = new SqlCommand("DELETE FROM [tblCNCBending] WHERE SubOA='" + hdnSubOANo.Value + "'", con);
+                    cmdDelete.ExecuteNonQuery();
+                }
+                else
+                {
+                    int inquity = Convert.ToInt32(Inwardqty) - Convert.ToInt32(TotalReturnInward);
+
+                    SqlCommand cmdsupdate = new SqlCommand("UPDATE [tblCNCBending] SET [InwardQty] = '" + inquity + "' WHERE SubOA='" + hdnSubOANo.Value + "'", con);
+                    cmdsupdate.ExecuteNonQuery();
+                }
             }
         }
         else
